Compile `export default <expr>` to `module.exportDefault(<expr>)`.
This allows us to simplify the module.export API significantly, since it
no longer sometimes takes a string as its first argument. Now `getters` is
always an object of methods that get the values of local variables, and
`constant` is an optional boolean to indicate that the getter functions
always return the same values, which helps with #134.