Call namespace setters once per namespace rather than once per name.
After a closer reading of the ECMAScript 2017 draft specification, I
recently realized that modules are allowed (even required) to reuse the
same (===) namespace whenever any other module uses the

  import * as ns from "module"

syntax. In other words, the `ns` object does not need to (and indeed
should not) be unique to each import declaration.

This is great news for Reify, because it means the generated code can
treat namespace objects the same as any other exported values, rather than
having to call "*" setters repeatedly for each name in the namespace.

Spec reference:
https://tc39.github.io/ecma262/#sec-getmodulenamespace
https://tc39.github.io/ecma262/#sec-modulenamespacecreate