Remove options.generateArrowFunctions (now always true).
The reason @jdalton and I originally provided this option was to avoid
generating arrow functions for getter and setter functions in Node 4,
since they are more costly than normal functions (in Node 4 at least).

Now that we're using method syntax for setter functions, which are called
much more often than getter functions, more than half the reason for this
option to exist has disappeared.

In Meteor, arrow functions will be compiled to normal functions in Node 4
anyway, so the simplification of removing this option seems relatively
free from negative consequences, all in all.