Guide AST traversal based on a RegExp scan of possible indexes.
When an AST traversal is looking for syntax that contains a known set of
identifiers, such as `import` and `export`, we can ignore entire subtrees
of the AST that do not contain any possible indexes of those identifiers,
for a potentially massive savings in traversal time.

The list of possible indexes can be computed quickly using a regular
expression, without actually parsing the code, since it's not the end of
the world if there are false positives in the possibleIndexes array.

Closes #113, finally! :tada: