Quote .export properties for compatibility with very old browsers.
Still seeing failures in PhantomJS because of this, unfortunately.

Generated code that calls module.export is safe because it can be compiled
with plugins like @babel/plugin-transform-property-literals, but the
reify/lib/runtime/entry.js code might not be compiled, either because it's
in node_modules or because it's part of the primordial module system
rather than application code.