Allow array-of-strings module.link setter values, too.
Since a single local variable/value can be exported as multiple different
export names (rare but possible), we need a shorthand for setting multiple
exports with one setter function.

With this commit, we no longer modify exports or module.exports anywhere
in generated code, which should make it easier to change the relationship
between the namespace object and module.exports in the future.